feat(interpretation): SJRA-1430 rework header and context
