+++ frontend/apps/case/src/entity/variants/interpretation/last-updated-banner.tsx
@@ -1,9 +1,8 @@
-import { format, formatDate } from 'date-fns';
+import { format } from 'date-fns';
 import { InfoIcon } from 'lucide-react';
 
 import { Alert, AlertDescription } from '@/components/base/shadcn/alert';
 import { useI18n } from '@/components/hooks/i18n';
-import DateTime from '@/components/base/date/datetime';
 
 type InterpretationLastUpdatedBannerProps = {
   updated_by_name?: string;
@@ -15,7 +14,7 @@ function InterpretationLastUpdatedBanner({ updated_by_name, updated_at }: Interp
   if (!updated_by_name || !updated_at) return null;
 
   return (
-    <Alert className="flex gap-2 items-center">
+    <Alert className="flex gap-2 items-center mb-3">
       <div>
         <InfoIcon size={16} />
       </div>
